prepare 3.2.0 release (#260)
## [3.2.0] - 2024-05-01
### Added:
- Added an optional timeout to the `waitForInitialization` method. When
a timeout is specified the returned promise will be rejected after the
timeout elapses if the client has not finished initializing within that
time. When no timeout is specified the returned promise will not be
resolved or rejected until the initialization either completes or fails.

### Changed:
- The track method now validates that the provided metricValue is a
number. If a metric value is provided, and it is not a number, then a
warning will be logged.

- Updated the link-dev script to better support multiple platforms.

### Fixed:
- Fixed the documentation for `evaluationReasons` for the `identify`
method.

---------

Co-authored-by: Rouel Joseph Soberano <rsoberano@launchdarkly.com>
Co-authored-by: Rouel Soberano <96156061+rsoberano-ld@users.noreply.github.com>
Co-authored-by: Yusinto Ngadiman <yusinto@gmail.com>
Co-authored-by: Matthew M. Keeler <mkeeler@launchdarkly.com>
Co-authored-by: LaunchDarklyReleaseBot <launchdarklyreleasebot@launchdarkly.com>
Co-authored-by: Ryan Lamb <4955475+kinyoklion@users.noreply.github.com>