fix: Fix a race condition when handling multiple flag updates. (#365)
The setState function is effectively asynchronous, multiple setState
methods may be called in a single cycle before this.state is updated.

So if there were multiple dispatches of changes within a single batch,
then some of the updates could be missed.

The way to handle this in react is to use the previous state provided
using the method version of setState. The operations will be done
sequentially with each having the correct view of the previous state.